Criado módulo ``utils.functions`` para melhor organização de funções que apoiam de alguma maneira as tarefas relacionadas a sincronização de dados
